+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### 2.11.1
 * added all property types to all classes;
 * updated for `php-tools` 1.7.4;
-* tests have been made compatible with Xampp on Windows.
+* tests have been made compatible with Xampp on Windows;
+* many, small improvements to the code and tests, also suggested by PhpStorm.
 
 ### 2.11.0
 * requires at least PHP 7.4;